pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "convaier"
version = "0.1.0"
description = "Local AI-powered CI/CD pipeline"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "ollama>=0.4",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
convaier = "convaier.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
